requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pydantic-settings
httpx
motor
python-jose[cryptography]
email-validator
bcrypt
